requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
faiss-cpu
PyPDF2==3.0.1
openai>=1.12.0
tiktoken
gunicorn